r27777 closed repo rev.nr.:   24382: Bugfix for erodable layers in NetCDF map file: Turned out to be wrong in our NetCDF writer, not in QUICKPLOT: variable ero had dimensions: time, nerolaydim, nFrac Coordinate x and y have dimensions: nNetNode (so not nerolaydim)
In NUMBERS these were the same:
nNetNode = 53118 ;
...
time = UNLIMITED ; // (1 currently)
nFrac = 1 ;
nerolaydim = 53118 ;

But CF does not allow that the dimension ID is different (even if the size/length is the same).
So: changed the id_erolaydim in our writer to either id_netnodedim, or id_flowelemdim, based on jaceneqtr==1 (using nf90_inq_dimid instead of nf90_def_dim).

NOTE: we still have a hidden bug left:
       ! TODO: AvD: size(grainlay,2) is always correct (mxn), but we have a problem if jaceneqtr==2 and mxn/=numk,
       ! because then the dimension for ero is set to nNetNode, and coordinate attribute refers to NetNode_x
       ! (both length numk), whereas ero itself is shorter than numk.