r2153 Updated two commits from Arjen Markus to the ds tree.
Change automatic array into an allocatable one to avoid stack overflows

Use OMP_GET_NUM_THREADS instead of OMP_GET_MAX_THREADS to get the number of threads that will be used.
(dlwqnf.f) Force the OpenMP-enabled loop to acknowledge the number of threads
that was previously set. This is apparently necessary in the context of OpenDA,
though it remains a riddle why.