r2825 Finally a proper fix for the file locking problems in update_version.cmd during simultanous builds.
Result: update_version.cmd can now run simultanously in same dir for
different projects; we now don't need to make projects dependent on each
other if that was only done to stage the call to update_version.cmd

Explanation:
a temp file that was created inside this batch file, gave
"The process cannot access[..] in use by another process",
because Visual Studio runs the Prebuild step for some projects at
the same time (e.g. the dll and exe projects for FM, or for 3Di).
Replaced the tempfile by a direct FOR /F .. IN ('command').