r3262 QUICKPLOT - major change file management.
Upon opening a file, QUICKPLOT reads (part of) the file contents. In many cases this is limited to the general file structure: index tables, data offsets, and header information to be able to quickly access the actual data later. However, for some files we read a bit more, possibly even the whole file (e.g. various Delft3D attribute files, and shipma xml files).

Now, what happens to that data? Since this data is needed to access the data for plots and export, we need to pass it around. Passing data around in MATLAB is usually not a problem since MATLAB will pass pointers around until you start modifying the data (lazy copy). However, this data is also referenced by the items plotted in a figure (for example to enable scrolling: the file may be closed in the QUICKPLOT main window, but you can still scroll through it in your figure -- unless you actually removed the file). Unfortunately, this file meta data is also saved when saving a figure as a .fig file. Usually this slightly increases the file size, but recent tests with shipma output show big .fig files and associated long saving time: a relatively simple figure could take half an hour to save and result in a .fig file of 70MB or bigger.

This new QUICKPLOT version splits the file meta data into the meta data for reconstruction (file names and options) and data for quick access (meta data read from the file). Only the former data is now passed around as proxy for the full meta data; the rest of the data is stored in a qp_proxy persistent variable and it will be retrieved only when entering the file specific reading routines. When a file is closed in the main window, the associated bulk meta data is deleted. When a figure event requests new access to the file, the meta data is reloaded (such meta data will not be deleted as long as QUICKPLOT remains open). As a result, the same shipma figure is stored in a few seconds and requires only 18kB.

Included in this code change are some routines to convert between relative and absolute paths. QUICKPLOT will now internally always use absolute paths to better be able to locate files for reloading. This was needed especially for the QUICKPLOT automated testbed which forces directory changes.