r36852 closed repo rev.nr.:   33457: Issue: unst-394 *re-merged xbeach, unfortunately the log history remains unavailable
NOTE: XBeach branch was created in #31417 (13-12-2013), but the files committed in the first following version #31593 (24-12-2013) were a local unversioned copy based on an older revision number, probably early december 2013.
In the merge of 10-04-14 (#33387), these accidental reverts were not committed back to the trunc, only the relevant (jawave.eq.4) were committed.