r39463 closed repo rev.nr.:   36068: Issue: UNST-606 targetArray in ec_gettimespacevalue works only if dataPtr *is* set during addtimespacerelation.
Functionality was fully operational. Only 1 line should be commented in meteo1.f90, not 3 (that deletes the entire case and skips connection to the target_item). In unstruc.F90, the ec_getValues call needs to get the target array name as final argument. Tested on "e02_dflowfm\f09_structures\c030_Weir_1D\".

As an example, this behaviour is set as enforced default for quantity=damlevel.