r5803 DELFT3D-34110: Add information on the line number where an error was discovered in DELWAQ1.
If a token of an unexpected type is encountered, the error message will now include the line number at which it was found. That does not mean that all error messages include the line number, but definitely the more annoying ones - that is, where a large amount of data is read without much structure, as per the use case.