r6394 Merging EC-module for FM with the modified EC-module used for channelflow. Preliminary, this version has still most tests failing in cf and 4 failing tests in FM under windows x64 (testbench), as we speak. Some of the other tests barely succeed, but show large differences. The common factor seems to be the ASTRONOMIC forcing (cmp as well as bc).
It causes NO detoriation of test results under Linux64 (Linux does not seem to be affected by the change of the EC-module), which suggests that
this has something to do with uninitialized memory. This is in line with the findings for CF: tests fail on the buildserver, but not when
carried out locally, while the external forcing coming from the EC-module seems to be EXACT. So the effect of the EC-module on CF is indirect.