r6555 Always use isequal before detailedcheck to improve performance.
Preallocate cells for inifile - when do we decide that a file isn't an ini-file: now we always accept the file, but error on its content (missing chapter/key). For large files this can be slow. Should there be a [ or = in the first X lines - what if somebody wants to comment out the initial block?