r7985 DELFT3D-35943: Use for limiting factors plot always "jet" for consistency. The colormap for limiting factors was always the (default) colormap of the figure. This used to be "jet" and since MATLAB 2014b is "parula". Since the pre-compiled QUICKPLOT has been based on 2013b, the most commonly experienced colormap is still "jet"; to keep the behaviour consistent when compiling with a newer MATLAB version, we are now making the behaviour more consistent across MATLAB versions.
Note there is still some version dependency since in recent MATLAB versions the colormap is an axes property. We are now setting the colormap of the limiting factors bars to "jet"; this avoids colormap conflicts in new MATLAB versions and falls back to changing the figure colormap in old MATLAB versions.

If you had created a figure with another colormap to which you added the limiting factors plot, the chosen colormap was used for the limiting factors bars, but this will now switch to using the "jet" colormap.