r9108 SOBEK3-1427: Remove directory https://svn.oss.deltares.nl/repos/delft3d/trunk/src/third_party_open/intel_fortran Re-added: third_party_open/intelredist containing IFORT18 dlls
The newest version of generic dlls will be placed in dest_share.
When compiling a kernel, the actual generic dlls are place in the kernels bin folder.
The DIMRset collector runs script "...\src\engines_gpl\dimr\scripts\dimr_artifacts.py", which removes duplicate dlls, assuming the version in dest_share must be kept, assuming the newest version of the dll can be used in combination with kernels build with older versions.