GRIDEDIT-588 Replaced std::vector with std::array
In a 2d mesh, for the edges-faces array, change the std::vector to a std::array
constrained with size 2. This change is ok, because each edge can only have, at
most, two neighbouring faces (boundary edges will have 1).

This results in fewer allocations of small (size 2) vectors and a small
performace improvement.