GRIDEDIT-588
Changed interface for IsEqual, the third parameter is now the relative tolerance
rather than a tolerance multiplier. This gives a more meaningful statement if
the value is passed and saves on a multiply in the function.

FindFacesRecursive
Replaced some emplace_back with push_back, this has a small improvement
on performance.

ComputeFaceCircumenter
Replaced vector with array, the maximum size is known at compile time.